move event handlers to core
