+++ packages/core/src/events/DefaultEventHandlers.ts
@@ -1,4 +1,3 @@
-import { isChromium, isLinux } from '@craftjs/utils';
 import isFunction from 'lodash/isFunction';
 import React from 'react';
 
@@ -7,6 +6,7 @@ import { Positioner } from './Positioner';
 import { createShadow } from './createShadow';
 
 import { Indicator, NodeId, DragTarget, NodeTree } from '../interfaces';
+import { isChromium, isLinux } from '../utils';
 
 export type DefaultEventHandlersOptions = {
   isMultiSelectEnabled: (e: MouseEvent) => boolean;
