+++ packages/core/src/events/EventHandlers/isEventBlockedByDescendant.ts
