+++ packages/core/src/events/index.tsx
@@ -1,4 +1,5 @@
 export * from './EventContext';
 export * from './CoreEventHandlers';
 export * from './DefaultEventHandlers';
+export * from './EventHandlers';
 export * from './Events';
