+++ packages/core/src/nodes/useInternalNode.ts
@@ -1,4 +1,3 @@
-import { wrapConnectorHooks } from '@craftjs/utils';
 import { useMemo, useContext } from 'react';
 import invariant from 'tiny-invariant';
 
@@ -6,6 +5,7 @@ import { NodeContext } from './NodeContext';
 
 import { ERROR_USE_NODE_OUTSIDE_OF_EDITOR_CONTEXT } from '../constants';
 import { useInternalEditor } from '../editor/useInternalEditor';
+import { wrapConnectorHooks } from '../events';
 import { Node } from '../interfaces';
 
 export function useInternalNode<S = null>(collect?: (node: Node) => S) {
