+++ packages/utils/src/index.ts
@@ -1,6 +1,5 @@
 export * from './useMethods';
 export * from './useCollector';
-export * from './EventHandlers';
 export * from './RenderIndicator';
 export * from './useEffectOnce';
 export * from './History';
